% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happiness_predictor.R
\name{happiness_predictor}
\alias{happiness_predictor}
\title{The prediction function for happiness level}
\usage{
happiness_predictor(a1, a2, a3, a4, a5, a6, a7)
}
\arguments{
\item{a1}{GDP_per_capital score, the extent to which a country's GDP per capita is above the world's lowest national average}

\item{a2}{Family score, the extent to which a country's social support is above the world's lowest national average}

\item{a3}{Life expectancy score, the extent to which a country's life expectancy is above the world's lowest national average}

\item{a4}{Freedom score, the extent to which a country's level of freedom is above the world's lowest national average}

\item{a5}{Trust in government score, the extent to which a country's trust in government is above the world's lowest national average}

\item{a6}{Generosity score, the extent to which a country's generosity level is above the world's lowest national average}

\item{a7}{Region that the country belongs to}
}
\description{
The prediction function for happiness level
}
\examples{
 \dontrun{
happiness_predictor(1.03192,1.23289,0.73608,0.37938,0.19090,0.11046,'Central and Eastern Europe')
}


}
